requirements.txt
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
torch==2.3.0
opencv-contrib-python==4.9.0.80
torchvision==0.18.0 
ipykernel==6.29.4